Update is_correct? to understand numbers as words
Use to_i method of String to convert the answer to an integer but only
if it matches the pattern /^-?\d+$/. This way, we only convert strings
to a number if it has a valid number format (one or more digits
following an optional negative sign). After converting the string to a
number, use humanize method to convert the number to its equivalent
english word.

Signed-off-by: Eugie Limpin <eugiellimpin@gmail.com>